+++ ee/src/main/java/org/jboss/as/ee/structure/EJBClientDescriptor12Parser.java
@@ -32,6 +32,7 @@ import java.util.Set;
 import javax.xml.stream.XMLStreamException;
 
 import org.jboss.as.ee.metadata.EJBClientDescriptorMetaData;
+import org.jboss.metadata.property.PropertyReplacer;
 import org.jboss.staxmapper.XMLExtendedStreamReader;
 
 /**
@@ -39,14 +40,14 @@ import org.jboss.staxmapper.XMLExtendedStreamReader;
  *
  * @author Jaikiran Pai
  * @author <a href="mailto:tadamski@redhat.com">Tomasz Adamski</a>
+ * @author <a href="mailto:wfink@redhat.com">Wolf-Dieter Fink</a>
  */
 class EJBClientDescriptor12Parser extends EJBClientDescriptor11Parser {
 
     public static final String NAMESPACE_1_2 = "urn:jboss:ejb-client:1.2";
 
-    public static final EJBClientDescriptor12Parser INSTANCE = new EJBClientDescriptor12Parser();
-
-    protected EJBClientDescriptor12Parser() {
+    protected EJBClientDescriptor12Parser(final PropertyReplacer propertyReplacer) {
+        super(propertyReplacer);
     }
 
     protected void parseClientContext(final XMLExtendedStreamReader reader,
@@ -101,55 +102,9 @@ class EJBClientDescriptor12Parser extends EJBClientDescriptor11Parser {
         unexpectedEndOfDocument(reader.getLocation());
     }
 
-    protected void parseEJBReceivers(final XMLExtendedStreamReader reader,
-            final EJBClientDescriptorMetaData ejbClientDescriptorMetaData) throws XMLStreamException {
-
-        // initialize the local-receiver-pass-by-value to the default true
-        Boolean localReceiverPassByValue = null;
-
-        final int count = reader.getAttributeCount();
-        for (int i = 0; i < count; i++) {
-            final EJBClientDescriptorXMLAttribute attribute = EJBClientDescriptorXMLAttribute.forName(reader
-                    .getAttributeLocalName(i));
-            final String val = reader.getAttributeValue(i);
-            switch (attribute) {
-                case EXCLUDE_LOCAL_RECEIVER:
-                    final boolean excludeLocalReceiver = Boolean.parseBoolean(val.trim());
-                    ejbClientDescriptorMetaData.setExcludeLocalReceiver(excludeLocalReceiver);
-                    break;
-                case LOCAL_RECEIVER_PASS_BY_VALUE:
-                    localReceiverPassByValue = Boolean.parseBoolean(val.trim());
-                    break;
-                default:
-                    unexpectedContent(reader);
-            }
-        }
-        // set the local receiver pass by value into the metadata
-        ejbClientDescriptorMetaData.setLocalReceiverPassByValue(localReceiverPassByValue);
-        while (reader.hasNext()) {
-            switch (reader.nextTag()) {
-                case END_ELEMENT: {
-                    return;
-                }
-                case START_ELEMENT: {
-                    final EJBClientDescriptorXMLElement element = EJBClientDescriptorXMLElement.forName(reader.getLocalName());
-                    switch (element) {
-                        case REMOTING_EJB_RECEIVER:
-                            this.parseRemotingReceiver(reader, ejbClientDescriptorMetaData);
-                            break;
-                        default:
-                            unexpectedElement(reader);
-                    }
-                    break;
-                }
-                default: {
-                    unexpectedContent(reader);
-                }
-            }
-        }
-        unexpectedEndOfDocument(reader.getLocation());
-    }
-
+    /**
+     * connectTimeout added
+     */
     protected void parseRemotingReceiver(final XMLExtendedStreamReader reader,
             final EJBClientDescriptorMetaData ejbClientDescriptorMetaData) throws XMLStreamException {
         String outboundConnectionRef = null;
@@ -162,14 +117,15 @@ class EJBClientDescriptor12Parser extends EJBClientDescriptor11Parser {
             final EJBClientDescriptorXMLAttribute attribute = EJBClientDescriptorXMLAttribute.forName(reader
                     .getAttributeLocalName(i));
             required.remove(attribute);
+            final String value = readResolveValue(reader, i);
             switch (attribute) {
                 case OUTBOUND_CONNECTION_REF:
-                    outboundConnectionRef = reader.getAttributeValue(i).trim();
+                    outboundConnectionRef = value;
                     remotingReceiverConfiguration = ejbClientDescriptorMetaData
                             .addRemotingReceiverConnectionRef(outboundConnectionRef);
                     break;
                 case CONNECT_TIMEOUT:
-                    connectTimeout = reader.getLongAttributeValue(i);
+                    connectTimeout = Long.parseLong(value);
                     break;
                 default:
                     unexpectedContent(reader);
